requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0

python-dotenv>=1.0.1
pydantic>=2.7.0
pydantic[email]>=2.7.0

SQLAlchemy>=2.0.30
psycopg2-binary>=2.9.9

python-jose>=3.3.0
passlib[bcrypt]>=1.7.4

python-multipart>=0.0.9